backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
icalendar
google-auth
google-auth-oauthlib
google-api-python-client
pydantic
python-dotenv
requests
